fixed next/font load on ALL pages and toast.
